+++ ecosystem/typescript/sdk/CHANGELOG.md
@@ -3,10 +3,16 @@
 All notable changes to the Aptos Node SDK will be captured in this file. This changelog is written by hand for now. It adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+
+- Introduce and use `@aptos-labs/aptos-client` package to manage and handle the client used in the SDK
+
+## 1.13.3 (2023-07-19)
+
 - Add support for `null` or `undefined` as absense of an Option
 - Fix `Option<LongerType>` for input arguments in entry functions
 
 ## 1.13.2 (2023-07-12)
+
 - Add support for `Option<T>` and `vector<Object<T>>` in the SDK
 
 ## 1.13.1 (2023-07-06)